requirements.txt
# Django & Core
Django==5.0.1
python-decouple==3.8
psycopg2-binary==2.9.9
django-environ==0.11.2

# Django REST Framework
djangorestframework
django-filter
drf-spectacular

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-beat
django-celery-results==2.5.1

# Email
imapclient==2.3.1
email-validator==2.3.0

# AI Integration
anthropic==0.18.1

# Security & Auth
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
sentry-sdk==1.40.0

# Development
django-debug-toolbar==4.2.0
faker==22.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0